Fix C++ TsFile writer flush offsets for multi-GB files. (#839)
* Fix C++ TsFile writer flush offsets for multi-GB files.

Sync file_base_offset after flush, reset the write stream and consumer,
and add tree/table 4GB regression tests.

* Add Python 4GB large-file tests and fix MSVC test includes.

Mirror C++ tree/table ~4GB regression coverage in pytest and drop POSIX-only headers from large-file gtests.

* Merge large-file test helpers into conftest.py.

Keep shared constants and utilities in one pytest entry point for tree/table 4GB tests.

* Fix Windows large-file truncate and stat handling in C++.

* fix msvc compilation

* removed unused test

+++ cpp/src/file/write_file.cc
@@ -131,7 +131,7 @@ int WriteFile::close() {
 int WriteFile::truncate(int64_t size) {
     ASSERT(fd_ > 0);
 #ifdef _WIN32
-    if (_chsize_s(fd_, static_cast<long>(size)) != 0) {
+    if (_chsize_s(fd_, size) != 0) {
         return E_FILE_WRITE_ERR;
     }
 #else
